requirements.txt
# Core Dependencies
torch==2.8.0
torchaudio==2.8.0
faster-whisper==1.0.2
pyannote.audio==4.1.1

# Audio Processing
soundfile==0.12.1
librosa==0.10.0

# Utilities
numpy==1.26.4